requirements.txt
gradio==3.50.2
scikit-learn
numpy
pillow
